Refactor by renaming folder
