+++ backend/src/main/java/com/rhyun/backend/flight/repository/FlightRepository.java
@@ -1,6 +1,6 @@
-package com.rhyun.backend.flights.repository;
+package com.rhyun.backend.flight.repository;
 
-import com.rhyun.backend.flights.model.Flight;
+import com.rhyun.backend.flight.model.Flight;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
